brain/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
redis[asyncio]>=5.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
openai>=1.30.0
requests>=2.31.0
httpx>=0.27.0
apscheduler>=3.10.0
stripe>=9.0.0
